+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 Refactored:
   - `mappersmith`: Migrated `middleware` to typescript #306
 
+Added:
+  - `mappersmith`: Added service workers support with default gateway `fetch` #311
+
 ## 2.38.1
 
 Fixed: